backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Rate Limiting ──────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── ML Stack (Python 3.13 compatible) ─────────────────────────────────────
numpy>=2.1.0
scikit-learn>=1.5.2
xgboost>=2.1.0
shap>=0.46.0
imbalanced-learn>=0.12.3

# ── Data ──────────────────────────────────────────────────────────────────
pandas>=2.2.2

# ── URL / Domain Analysis ─────────────────────────────────────────────────
tldextract>=5.1.2
python-whois>=0.9.4
requests>=2.31.0

# ── Serialization ─────────────────────────────────────────────────────────
pydantic>=2.7.1
